split histories into 4 separate files
previously, the wish.txt file contained a dictionary which had a maximum of 10 million entries. When 10 million entries were achieved, the lag was very noticeable when making wishes. The program would try to save everything into one file. Now, after splitting them into 4 files, the dictionary still exists, but now the maximum amount of wishes saved to a file is 2.5m down from 10m which is noticeably faster and makes for a much more seamless experience!